adds <pageoptions/>
